+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Annotation `alpha.aws.giantswarm.io/metadata-v2` to enable AWS Metadata API v2
+- Annotation `alpha.aws.giantswarm.io/aws-subnet-size` to customize subnet size of Control Plane and Node Pools
 - Annotation `alpha.aws.giantswarm.io/update-max-batch-size` to configure max batch size in ASG update policy on cluster or machine deployment CR.
 - Annotation `alpha.aws.giantswarm.io/update-pause-time` to configure pause between batches in ASG update on cluster or machine deployment CR.
 
@@ -22,8 +24,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Annotation `alpha.giantswarm.io/aws-metadata-v2` to enable AWS Metadata API v2
 - Add `terminate-unhealthy-node` feature to automatically terminate bad and
   unhealthy nodes in a Cluster.
-- Add `alpha.giantswarm.io/aws-metadata-v2` annotation to enable AWS Metadata
-  API v2.
 
 ### Fixed